pyproject.toml
[tool.poetry]
name = "iabank-backend"
version = "0.1.0"
description = "Fundação backend IABank alinhada ao blueprint arquitetural."
authors = ["IABank Platform Team <platform@iabank.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "4.2.26"
djangorestframework = "3.15.2"
django-prometheus = "2.3.1"
structlog = "24.1.0"
PyYAML = "6.0.2"
jsonschema = "4.23.0"
django-ratelimit = "3.0.1"
sentry-sdk = { version = "^2.8.0", extras = ["django"] }
psycopg = {version = "3.1.19", extras = ["binary"]}
celery = "5.3.6"
redis = "5.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.3"
pytest-django = "4.8.0"
bandit = "1.7.9"
ruff = "0.6.9"
pytest-cov = "5.0.0"
radon = "6.0.1"
pre-commit = "^3.5.0"
factory-boy = "3.3.0"

[tool.ruff]
extend-exclude = [
  "documentacao_oficial_spec-kit/**",
]

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
